♻️ refactor: 五维评分器替代旧 score
