+++ dragon_quant/orchestrator.py
@@ -2,11 +2,11 @@
 Orchestrator — 编排层主流程
 
 阶段：
-  A. 板块排行 → 前10涨/前10跌
-  B. 候选股筛选 → 每板块按5日累计涨幅取前5（过滤ST/双创/北交所） + 多概念跟踪
-  C. 连板高度 + 排序（连板优先） → top_n
-  D. 并发加载评分数据 → 全部写共享缓存
-  E. 主进程四维打分 → 只对 top_n 评分
+  A. 板块排行 → 前5涨/前20跌
+  B. 候选股筛选 → 每个领涨行业取当日全部涨停股（过滤ST/双创/北交所）
+  C. 连板高度 + 排序（连板优先） → 候选池全部评分
+  D. 并发加载五维评分数据 → 全部写共享缓存
+  E. 主进程五维打分
   F. 输出报告（表格 + 自然语言 + 持久化）
 
 对外接口：
@@ -15,12 +15,10 @@ Orchestrator — 编排层主流程
 """
 
 import json
-import os
 import sys
 import time
 import functools
 from datetime import datetime, timezone, timedelta
-from pathlib import Path
 from typing import Optional
 
 from dragon_quant.models.types import Candidate, StockInfo
@@ -29,7 +27,7 @@ from dragon_quant.rate_limit import RateLimiter
 from dragon_quant.providers import create_providers
 from dragon_quant.logging.logger import ScanLogger
 from dragon_quant.logging.reporter import ReportBuilder
-from dragon_quant.storage.paths import DATA_DIR, SHARED_DIR, RESULTS_DIR
+from dragon_quant.storage.paths import RESULTS_DIR
 from dragon_quant._version import __version__
 
 STATISTICAL_CONCEPT_PREFIXES = (
@@ -39,8 +37,7 @@ STATISTICAL_CONCEPT_PREFIXES = (
     "最近多板", "东方财富热股",
 )
 
-RANK_UP_COUNT = 8     # 领涨板块取前 8（候选筛选 + 5分K）
-RANK_UP_COUNT_V2 = 5  # scorers_v2：领涨板块取前 5（候选为板块内当日涨停股）
+RANK_UP_COUNT = 5     # 领涨行业取前 5（候选为板块内当日涨停股）
 RANK_DOWN_COUNT = 20  # 领跌板块取前 20（资金承接 + 5分K）
 
 # 数据源 / 接口中文名（仅用于控制台失败提示）
@@ -175,7 +172,7 @@ def _compute_5day_return(klines: list) -> float:
 
 
 def _to_serializable(obj):
-    """递归转换 dataclass 为 dict"""
+    """递归转换 dataclass / bytes / 容器为 JSON 友好结构。"""
     if hasattr(obj, '__dataclass_fields__'):
         return {
             f.name: _to_serializable(getattr(obj, f.name))
@@ -185,29 +182,9 @@ def _to_serializable(obj):
         return [_to_serializable(i) for i in obj]
     if isinstance(obj, dict):
         return {k: _to_serializable(v) for k, v in obj.items()}
-    # Convert bytes/timestamp objects that might cause issues
     if isinstance(obj, bytes):
         return obj.decode('utf-8', errors='replace')
     return obj
-
-
-def _save_shared(cache: DataCache):
-    """导出缓存快照到共享文件，子进程读取"""
-    timestamp = time.strftime("%Y%m%d_%H%M%S")
-    SHARED_DIR.mkdir(parents=True, exist_ok=True)
-    path = SHARED_DIR / f"dq_shared_{timestamp}.json"
-    snapshot = cache.snapshot()
-    serializable = _to_serializable(snapshot)
-    with open(path, "w") as f:
-        json.dump(serializable, f, ensure_ascii=False)
-    return str(path)
-
-
-def _load_shared(path: str, cache: DataCache):
-    """从共享文件恢复缓存"""
-    with open(path) as f:
-        data = json.load(f)
-    cache.load_snapshot(data)
 
 
 # ════════════════════════════════════════════════════════════
@@ -219,60 +196,8 @@ def _score_one(cand: Candidate, cache: DataCache,
                all_sector_codes: list[str],
                sector_name_map: dict[str, str],
                logger) -> dict:
-    from dragon_quant.scorers.drive import score as score_drive
-    from dragon_quant.scorers.anti_drop import score as score_anti_drop
-    from dragon_quant.scorers.leadership import score as score_leadership
-    from dragon_quant.scorers.absorption import score as score_absorption
-
-    scorers = [
-        ("drive",      score_drive,      0.35),
-        ("anti_drop",  score_anti_drop,  0.15),
-        ("leadership", score_leadership, 0.25),
-        ("absorption", score_absorption, 0.25),
-    ]
-
-    dims = {}
-    composite = 0.0
-    for dim_name, score_fn, weight in scorers:
-        try:
-            kwargs = {"code": cand.code, "cache": cache}
-            if dim_name == "drive":
-                kwargs["candidate_pool"] = candidate_pool
-            if dim_name in ("drive", "leadership", "absorption"):
-                kwargs["primary_sector"] = cand.primary_sector
-            if dim_name == "absorption":
-                kwargs["all_sector_codes"] = all_sector_codes
-                kwargs["sector_name_map"] = sector_name_map
-
-            sr = score_fn(**kwargs)
-            dims[dim_name] = {"score": sr.score, "weight": sr.weight, "details": sr.details}
-            composite += sr.score * sr.weight
-
-            # 结构化日志
-            logger.scorer(dim_name, cand.code, score=sr.score, weight=sr.weight,
-                          **sr.details)
-        except Exception as e:
-            logger.error(f"scorer:{dim_name}", f"打分异常: {e}",
-                         code=cand.code, exception=str(e))
-            dims[dim_name] = {"score": 50.0, "weight": weight, "details": {"error": str(e)}}
-            composite += 50.0 * weight
-
-    return {
-        "code": cand.code, "name": cand.name,
-        "concepts": cand.concepts, "board_count": cand.board_count,
-        "primary_sector": cand.primary_sector,
-        "primary_sector_name": sector_name_map.get(cand.primary_sector, ""),
-        "composite_score": round(composite, 2), "dimensions": dims,
-    }
-
-
-def _score_one_v2(cand: Candidate, cache: DataCache,
-                  candidate_pool: list[Candidate],
-                  all_sector_codes: list[str],
-                  sector_name_map: dict[str, str],
-                  logger) -> dict:
-    """scorers_v2 五维评分 + 门槛/加权聚合（输出兼容 _score_one 结构）。"""
-    from dragon_quant.scorers_v2 import evaluate
+    """五维评分 + 门槛/加权聚合。"""
+    from dragon_quant.scorers import evaluate
 
     verdict = evaluate(
         cand.code, cache,
@@ -303,16 +228,19 @@ def _print_cached_output(output_data: dict, top_n: int):
     print(f"\n{'═'*56}")
     print(f"🐉 龙头战法扫描完成 (缓存读取) - {output_data.get('timestamp', '?')}")
     print(f"{'═'*56}")
-    print(f"\n{'代码':8s} {'名称':8s} {'综合':>6s}  {'带动':>6s}  {'抗跌':>6s}  {'领涨':>6s}  {'承接':>6s}")
-    print("-" * 56)
+    print(f"\n{'代码':8s} {'名称':8s} {'综合':>6s}  {'带动':>6s}  {'领涨':>6s}  {'抗跌':>6s}  {'流动':>6s}  {'承接':>6s}  真龙")
+    print("-" * 64)
     display_list = output_data.get("ranking", [])[:top_n]
     for r in display_list:
+        dims = r.get("dimensions", {})
+        mark = "🐉" if r.get("is_true_dragon") else "✗"
         print(f"{r['code']:8s} {r.get('name', ''):8s} "
               f"{r.get('composite_score', 0):6.1f}  "
-              f"{r.get('dimensions', {}).get('drive', {}).get('score', 0):6.1f}  "
-              f"{r.get('dimensions', {}).get('anti_drop', {}).get('score', 0):6.1f}  "
-              f"{r.get('dimensions', {}).get('leadership', {}).get('score', 0):6.1f}  "
-              f"{r.get('dimensions', {}).get('absorption', {}).get('score', 0):6.1f}")
+              f"{dims.get('drive', {}).get('score', 0):6.1f}  "
+              f"{dims.get('leadership', {}).get('score', 0):6.1f}  "
+              f"{dims.get('anti_drop', {}).get('score', 0):6.1f}  "
+              f"{dims.get('liquidity', {}).get('score', 0):6.1f}  "
+              f"{dims.get('absorption', {}).get('score', 0):6.1f}   {mark}")
     print(f"\n{'═'*56}")
     print(f"📋 完整详细报告 (来自缓存)")
     print(f"{'═'*56}")
@@ -357,7 +285,7 @@ def _get_trade_date() -> Optional[str]:
 
 def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
          verbose: bool = True, force: bool = False,
-         scorers: str = "v1", refresh_provider_cache: bool = False) -> dict:
+         scorers: str = "v2", refresh_provider_cache: bool = False) -> dict:
     """龙头战法完整扫描 — Programmtic API
 
     返回结构化 dict:
@@ -366,14 +294,15 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
         "elapsed_s": float,
         "params": {"top_n", "candidates_n", "workers"},
         "sectors": {"up": [...], "down": [...]},
-        "ranking": [  ← 仅 top_n 只，按综合分降序
+        "ranking": [  ← 全部评分结果，按综合分降序
           {
             "code", "name", "concepts", "board_count",
             "primary_sector", "primary_sector_name",
             "composite_score", "dimensions": {
               "drive": {"score", "weight", "details"},
-              "anti_drop": {...},
               "leadership": {...},
+              "anti_drop": {...},
+              "liquidity": {...},
               "absorption": {...},
             }
           }
@@ -386,16 +315,15 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
     """
     bj_tz = timezone(timedelta(hours=8))
     now = datetime.now(bj_tz)
-    use_v2 = (scorers == "v2")
-    source = "v2" if use_v2 else "v1"
-    rank_up_count = RANK_UP_COUNT_V2 if use_v2 else RANK_UP_COUNT
+    source = "v2"
+    rank_up_count = RANK_UP_COUNT
     
     # 1. 交易时间段拦截 (工作日的 9:00 - 15:00)
     is_trading_day = now.weekday() < 5
     if not force and is_trading_day and 9 <= now.hour < 15:
         msg = (
             "⚠️ 拦截: 当前处于 A 股交易时段 (尚未收盘)！\n"
-            "【原因】龙头战法的四维打分强依赖于【日K线最终收盘价】、【最终连板高度】和【全天资金承接情况】。\n"
+            "【原因】龙头战法的五维打分强依赖于【日K线最终收盘价】、【最终连板高度】和【全天资金承接情况】。\n"
             "在 15:00 收盘前，行情数据处于剧烈波动中，提前执行会导致评分严重失真、排名不准，极易选出伪龙头。\n"
             "💡 强烈推荐在 15:00 收盘后再执行 scan 命令获取准确结果。\n"
             "（若为盘中测试需要，请使用 --force 参数强制执行）"
@@ -442,7 +370,6 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
     logger = ScanLogger()
 
     providers = create_providers(logger=logger)
-    em = providers["eastmoney"]
     ths = providers["ths"]
     xq = providers["xueqiu"]
     tx = providers["tencent"]
@@ -492,8 +419,8 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
             return False
         return True
 
-    # 一次抓取全部概念排行（provider 内已多页+本地排序），本地切涨幅/跌幅榜，
-    # 避免重复抓 8 页（desc/asc 数据相同）。缓存原始榜，黑名单过滤仍在内存做，
+    # 一次抓取全部行业排行（provider 内已多页+本地排序），本地切涨幅/跌幅榜。
+    # 缓存原始榜，黑名单过滤仍在内存做，
     # 保证黑名单改动即时生效。
     ranking_raw = _cf_sync("sector:ranking",
                            lambda: ths.get_sector_ranking(asc=False)) or []
@@ -524,12 +451,11 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
     sector_components: dict[str, list[StockInfo]] = {}
     sector_filtered: dict[str, list[StockInfo]] = {}
 
-    # 提交前10涨板块成分股请求（过 RateLimiter 防 burst 反爬）
-    #   v2: all_pages=True 翻页到 ≈50 只，供 leadership 涨幅分位用更大样本
+    # 提交领涨板块成分股请求（过 RateLimiter 防 burst 反爬）
     for s in top10_up:
         _cf("ths", "ths", f"sector:components:{s.code}",
             (lambda sc=s.code: ths.get_sector_components(sc, page=1,
-                                                         all_pages=use_v2)),
+                                                         all_pages=True)),
             namespace=source)
     limiter.wait_all()
     fail_seen = _report_api_failures(logger, fail_seen, verbose)
@@ -540,17 +466,11 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
         filtered = [c for c in components if _is_valid_candidate(c)]
         sector_filtered[s.code] = filtered
 
-    # 收集需要拉K线的唯一股票
-    #   v1: 每板块前 pre_n 只（按成分股原序）
-    #   v2: 每板块当日所有涨停个股（pct ≥ LIMIT_UP_PCT）
+    # 收集需要拉K线的唯一股票：每板块当日所有涨停个股（pct ≥ LIMIT_UP_PCT）。
     unique_codes: dict[str, StockInfo] = {}
-    pre_n = max(candidates_n * 2, 10)
     LIMIT_UP_PCT = 9.9
     for s in top10_up:
-        if use_v2:
-            picks = [st for st in sector_filtered[s.code] if st.pct >= LIMIT_UP_PCT]
-        else:
-            picks = sector_filtered[s.code][:pre_n]
+        picks = [st for st in sector_filtered[s.code] if st.pct >= LIMIT_UP_PCT]
         for stock in picks:
             if stock.code not in unique_codes:
                 unique_codes[stock.code] = stock
@@ -569,19 +489,13 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
         klines = cache.get(f"kline:day:{code}") or []
         stock.five_day_return = _compute_5day_return(klines)
 
-    # 候选构建：
-    #   v1: 每板块按5日累计涨幅重排取前 candidates_n
-    #   v2: 取每板块当日所有涨停个股（不截断）
+    # 候选构建：取每板块当日所有涨停个股（不截断）。
     for s in top10_up:
         pre_list = [st for st in sector_filtered[s.code] if st.code in unique_codes]
-        if use_v2:
-            top_n_stocks = [st for st in pre_list if st.pct >= LIMIT_UP_PCT]
-            if verbose:
-                print(f"   [{s.name}] 涨停候选 {len(top_n_stocks)} 只"
-                      f"（成分股{len(sector_filtered[s.code])}只）")
-        else:
-            pre_list.sort(key=lambda c: c.five_day_return, reverse=True)
-            top_n_stocks = pre_list[:candidates_n]
+        top_n_stocks = [st for st in pre_list if st.pct >= LIMIT_UP_PCT]
+        if verbose:
+            print(f"   [{s.name}] 涨停候选 {len(top_n_stocks)} 只"
+                  f"（成分股{len(sector_filtered[s.code])}只）")
         for stock in top_n_stocks:
             if stock.code in all_candidates:
                 all_candidates[stock.code].concepts.append(s.name)
@@ -619,14 +533,8 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
     for c in candidate_pool:
         kline = cache.get(f"kline:day:{c.code}") or []
         c.board_count = _compute_consecutive_boards(kline)
-        # v2: 5日总涨幅写入候选（leadership 涨幅分位用），复用日K
-        if use_v2:
-            c.fived_pct = _compute_5day_return(kline)
-
-    # 拉大盘日K线（上证指数 000001 的 K 线用于跳水日检测）
-    market_code = "000001"
-    market_kline = _cf_sync(f"kline:day:{market_code}",
-                            lambda: xq.get_kline(market_code, days=30)) or []
+        # 5日总涨幅写入候选（leadership 涨幅分位用），复用日K。
+        c.fived_pct = _compute_5day_return(kline)
 
     # 排序（连板优先）后对候选池全部个股评分，不再截断
     candidate_pool.sort(key=lambda c: (c.board_count, len(c.concepts)), reverse=True)
@@ -646,38 +554,23 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
     if verbose:
         print("⏳ Phase D — 并发加载")
 
-    # T1: 板块5分K
-    #   v1: 当日5分K（provider 内聚合）
-    #   v2: 近10交易日历史5分K（虹吸回看），并额外拉主板块当日1分K（带动/抗跌基准）
-    #   口径不同 → 按 source 隔离缓存文件，避免 v1/v2 串味
+    # T1: 板块历史5分K（虹吸回看）+ 主板块当日1分K（带动/抗跌基准）
     all_sectors = top10_up + top10_down
     for s in all_sectors:
-        if use_v2:
-            _cf("ths", "ths", f"kline:5min:sector:{s.code}",
-                (lambda sc=s.code: ths.get_sector_5min_kline_history(sc, days=10)),
-                namespace=source)
-        else:
-            _cf("ths", "ths", f"kline:5min:sector:{s.code}",
-                (lambda sc=s.code: ths.get_sector_5min_kline(sc)),
-                namespace=source)
-    if use_v2:
-        # 主板块当日1分K（领涨板块即可，带动/板块抗跌基准）
-        for s in top10_up:
-            _cf("ths", "ths", f"kline:1min:sector:{s.code}",
-                (lambda sc=s.code: ths.get_sector_1min_kline(sc)),
-                namespace=source)
+        _cf("ths", "ths", f"kline:5min:sector:{s.code}",
+            (lambda sc=s.code: ths.get_sector_5min_kline_history(sc, days=10)),
+            namespace=source)
+    for s in top10_up:
+        _cf("ths", "ths", f"kline:1min:sector:{s.code}",
+            (lambda sc=s.code: ths.get_sector_1min_kline(sc)),
+            namespace=source)
 
-    # T2: 候选股分时K线
-    #   v1: 只拉 ranking（top_n 评分池）
-    #   v2: 拉全部候选（均为涨停股，含 drive 封板池对比所需的同板块涨停股）
-    minute_targets = candidate_pool if use_v2 else ranking
-    for r in minute_targets:
+    # T2: 全部候选股分时K线（含 drive 封板池对比所需的同板块涨停股）
+    for r in candidate_pool:
         _cf("xueqiu", "minute_kline", f"kline:1min:{r.code}",
             lambda c=r.code: xq.get_minute_kline(c))
-    if use_v2:
-        # 大盘当日1分K（抗跌性大盘基准）
-        _cf("xueqiu", "minute_kline", "kline:1min:000001",
-            lambda: xq.get_minute_kline("000001"))
+    _cf("xueqiu", "minute_kline", "kline:1min:000001",
+        lambda: xq.get_minute_kline("000001"))
 
     # T3: 腾讯批量行情（含收盘盘口 bid1/ask1，liquidity 封单用）
     all_codes = set()
@@ -699,10 +592,10 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
             print(f"   ⚠️ 数据加载完成（部分接口失败，详见上方）")
 
     # ────────────────────────────────────────────
-    # Phase E: 主进程打分 — 只对 top_n 只
+    # Phase E: 主进程打分 — 对候选池全部个股
     # ────────────────────────────────────────────
     if verbose:
-        print("🔨 Phase E — 五维打分" if use_v2 else "🔨 Phase E — 四维打分")
+        print("🔨 Phase E — 五维打分")
 
     # 注入元数据到缓存（供 scorer 读取）
     cache.set("__meta__:candidates", [
@@ -716,37 +609,25 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
     sector_name_map = {s.code: s.name for s in all_sectors}
     cache.set("__meta__:sector_name_map", sector_name_map)
 
-    # v2 用全候选池做板块内对比（连板/涨幅分位、封板池）；v1 用 ranking
-    peer_pool = candidate_pool if use_v2 else ranking
+    # 用全候选池做板块内对比（连板/涨幅分位、封板池）。
+    peer_pool = candidate_pool
     results = []
     for cand in ranking:
         try:
-            if use_v2:
-                sr = _score_one_v2(cand, cache, peer_pool,
-                                   [s.code for s in top10_down],
-                                   sector_name_map, logger)
-            else:
-                sr = _score_one(cand, cache, ranking, [s.code for s in top10_down],
-                               sector_name_map, logger)
+            sr = _score_one(cand, cache, peer_pool,
+                            [s.code for s in top10_down],
+                            sector_name_map, logger)
             results.append(sr)
             if verbose:
                 dims = sr.get("dimensions", {})
-                if use_v2:
-                    print(f"  {cand.code} {cand.name:6s}  "
-                          f"综合={sr['composite_score']:5.1f}  "
-                          f"{'✓真龙' if sr.get('is_true_dragon') else '✗'}  "
-                          f"带动={dims.get('drive',{}).get('score',0):5.1f}  "
-                          f"领涨={dims.get('leadership',{}).get('score',0):5.1f}  "
-                          f"抗跌={dims.get('anti_drop',{}).get('score',0):5.1f}  "
-                          f"流动={dims.get('liquidity',{}).get('score',0):5.1f}  "
-                          f"承接={dims.get('absorption',{}).get('score',0):5.1f}")
-                else:
-                    print(f"  {cand.code} {cand.name:6s}  "
-                          f"综合={sr['composite_score']:5.1f}  "
-                          f"带动={dims.get('drive',{}).get('score',0):5.1f}  "
-                          f"抗跌={dims.get('anti_drop',{}).get('score',0):5.1f}  "
-                          f"领涨={dims.get('leadership',{}).get('score',0):5.1f}  "
-                          f"承接={dims.get('absorption',{}).get('score',0):5.1f}")
+                print(f"  {cand.code} {cand.name:6s}  "
+                      f"综合={sr['composite_score']:5.1f}  "
+                      f"{'✓真龙' if sr.get('is_true_dragon') else '✗'}  "
+                      f"带动={dims.get('drive',{}).get('score',0):5.1f}  "
+                      f"领涨={dims.get('leadership',{}).get('score',0):5.1f}  "
+                      f"抗跌={dims.get('anti_drop',{}).get('score',0):5.1f}  "
+                      f"流动={dims.get('liquidity',{}).get('score',0):5.1f}  "
+                      f"承接={dims.get('absorption',{}).get('score',0):5.1f}")
         except Exception as e:
             if verbose:
                 print(f"  ⚠️ {cand.code} 打分失败: {e}", file=sys.stderr)
@@ -779,24 +660,15 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
         # 提前初始化 Reporter，并为所有结果生成报告，存入 r["report_text"] 以便持久化
         reporter = ReportBuilder(logger)
         for r in results:
-            if use_v2:
-                r["report_text"] = reporter.build_stock_report_v2(
-                    r["code"], r.get("name", ""),
-                    r.get("board_count", 0), r.get("concepts", []),
-                    composite_score=r.get("composite_score", 0),
-                    dimensions=r.get("dimensions", {}),
-                    primary_sector_name=r.get("primary_sector_name", ""),
-                    is_true_dragon=r.get("is_true_dragon", False),
-                    reject_reason=r.get("reject_reason"),
-                )
-            else:
-                r["report_text"] = reporter.build_stock_report(
-                    r["code"], r.get("name", ""),
-                    r.get("board_count", 0), r.get("concepts", []),
-                    composite_score=r.get("composite_score", 0),
-                    dimensions=r.get("dimensions", {}),
-                    primary_sector_name=r.get("primary_sector_name", ""),
-                )
+            r["report_text"] = reporter.build_stock_report(
+                r["code"], r.get("name", ""),
+                r.get("board_count", 0), r.get("concepts", []),
+                composite_score=r.get("composite_score", 0),
+                dimensions=r.get("dimensions", {}),
+                primary_sector_name=r.get("primary_sector_name", ""),
+                is_true_dragon=r.get("is_true_dragon", False),
+                reject_reason=r.get("reject_reason"),
+            )
 
         output["ranking"] = results  # 返回全部评分结果
 
@@ -811,30 +683,18 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
             print(f"\n{'═'*56}")
             print(f"🐉 龙头战法扫描完成 ({elapsed:.0f}s)")
             print(f"{'═'*56}")
-            if use_v2:
-                print(f"\n{'代码':8s} {'名称':8s} {'综合':>6s}  {'带动':>6s}  {'领涨':>6s}  {'抗跌':>6s}  {'流动':>6s}  {'承接':>6s}  真龙")
-                print("-" * 64)
-                for r in display_list:
-                    dims = r.get("dimensions", {})
-                    mark = "🐉" if r.get("is_true_dragon") else "✗"
-                    print(f"{r['code']:8s} {r.get('name', ''):8s} "
-                          f"{r['composite_score']:6.1f}  "
-                          f"{dims.get('drive', {}).get('score', 0):6.1f}  "
-                          f"{dims.get('leadership', {}).get('score', 0):6.1f}  "
-                          f"{dims.get('anti_drop', {}).get('score', 0):6.1f}  "
-                          f"{dims.get('liquidity', {}).get('score', 0):6.1f}  "
-                          f"{dims.get('absorption', {}).get('score', 0):6.1f}   {mark}")
-            else:
-                print(f"\n{'代码':8s} {'名称':8s} {'综合':>6s}  {'带动':>6s}  {'抗跌':>6s}  {'领涨':>6s}  {'承接':>6s}")
-                print("-" * 56)
-                for r in display_list:
-                    dims = r.get("dimensions", {})
-                    print(f"{r['code']:8s} {r.get('name', ''):8s} "
-                          f"{r['composite_score']:6.1f}  "
-                          f"{dims.get('drive', {}).get('score', 0):6.1f}  "
-                          f"{dims.get('anti_drop', {}).get('score', 0):6.1f}  "
-                          f"{dims.get('leadership', {}).get('score', 0):6.1f}  "
-                          f"{dims.get('absorption', {}).get('score', 0):6.1f}")
+            print(f"\n{'代码':8s} {'名称':8s} {'综合':>6s}  {'带动':>6s}  {'领涨':>6s}  {'抗跌':>6s}  {'流动':>6s}  {'承接':>6s}  真龙")
+            print("-" * 64)
+            for r in display_list:
+                dims = r.get("dimensions", {})
+                mark = "🐉" if r.get("is_true_dragon") else "✗"
+                print(f"{r['code']:8s} {r.get('name', ''):8s} "
+                      f"{r['composite_score']:6.1f}  "
+                      f"{dims.get('drive', {}).get('score', 0):6.1f}  "
+                      f"{dims.get('leadership', {}).get('score', 0):6.1f}  "
+                      f"{dims.get('anti_drop', {}).get('score', 0):6.1f}  "
+                      f"{dims.get('liquidity', {}).get('score', 0):6.1f}  "
+                      f"{dims.get('absorption', {}).get('score', 0):6.1f}   {mark}")
 
             print(f"\n{'═'*56}")
             print(f"📋 完整详细报告")
@@ -845,7 +705,7 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
 
         # ── 持久化 ──
         timestamp = output["timestamp"]
-        scan_id = f"{source}_{timestamp[:8]}_{top_n}"  # v1_20260519_25 / v2_20260519_25
+        scan_id = f"{source}_{timestamp[:8]}_{top_n}"  # v2_20260519_25
 
         # 结构化日志 → SQLite
         try:
@@ -861,10 +721,7 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
         # 报告文本
         report_path = RESULTS_DIR / f"scan_report_{source}_{timestamp}.txt"
         with open(report_path, "w") as f:
-            if use_v2:
-                f.write(reporter.build_summary_report_v2(display_list))
-            else:
-                f.write(reporter.build_summary_report(display_list))
+            f.write(reporter.build_summary_report(display_list))
             f.write("\n\n")
             f.write(output["report_text"])
         output["report_path"] = str(report_path)
@@ -906,7 +763,7 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
 
                 # 5 日去重：该 code 上次入选距今 < 5 个交易日
                 #   仅对「跨日」(last_date 严格早于当天) 重复入选去重；
-                #   v1/v2 使用独立 dragons 表，因此同日另一评分器版本不会影响当前体系。
+                #   当前主流程固定使用 dragons_v2，兼容历史 v2 数据。
                 last_info = db.get_last_entry_with_rank(code, source=source)
                 if last_info:
                     last_date, old_rank = last_info
@@ -958,7 +815,7 @@ def scan(top_n: int = 5, candidates_n: int = 5, workers: int = 2,
 
 
 def run_scan(top_n: int = 25, candidates_n: int = 5, workers: int = 2,
-             force: bool = False, scorers: str = "v1",
+             force: bool = False, scorers: str = "v2",
              refresh_provider_cache: bool = False):
     """CLI 入口 — 同 scan() 但带 verbose 输出"""
     scan(top_n=top_n, candidates_n=candidates_n, workers=workers,
