+++ dragon_quant/scorers/drive.py
@@ -1,443 +1,278 @@
-"""
-带动性 scorer（权重 35%）
+"""scorers.drive — 带动性 (30%)。
 
-核心问题：这只股票封板时，板块里其他股票跟不跟？
-数据策略：在 Phase B 已获取的候选股范围内按板块分组比对封板时间。
+txt："它一涨停整个板块的小弟全都跟风高潮 / 观察哪支股票最先涨停并稳定封死，
+同时观察同板块个股是否能跟随他拉升"。仅当日盘面。
+三个子因子各满分100：封板最早 EARLY_W .40 + 带动板块 LEAD_W .35 + 板块共鸣 VOICE_W .25。
 """
+
+from __future__ import annotations
 
 from datetime import datetime
 from typing import Optional
-from dragon_quant.models.types import ScoreResult, KBar, StockInfo, Candidate
+
 from dragon_quant.cache.data_cache import DataCache
+from dragon_quant.models.types import Candidate, KBar, Quote, ScoreResult, StockInfo
+from dragon_quant.scorers import registry as R
+from dragon_quant.scorers.base import clip, common_minute_axis, gain_curve
+
+DIM = "drive"
+WEIGHT = R.DIM_WEIGHTS[DIM]
+EPS = 1e-9
 
 
-DRIVE_SAMPLE_LIMIT = 50
-
-
-def score(code: str, cache: DataCache, candidate_pool: Optional[list[Candidate]] = None,
-          primary_sector: str = "") -> ScoreResult:
-    """
-    Args:
-        code: 股票代码
-        cache: 共享数据缓存
-        candidate_pool: 全部候选股列表（用于板块内封板比对）
-        primary_sector: 候选股主板块代码
-    Returns:
-        ScoreResult(dim="drive", score, weight=0.35, details)
-    """
-    # 加载个股数据
-    stock_klines: list[KBar] = cache.get(f"kline:day:{code}") or []
-    stock_1min: list[KBar] = cache.get(f"kline:1min:{code}") or []
+def score(code: str, cache: DataCache, primary_sector: str = "",
+          candidate_pool: Optional[list[Candidate]] = None,
+          **kwargs) -> ScoreResult:
+    minute: list[KBar] = cache.get(f"kline:1min:{code}") or []
+    sector_min: list[KBar] = cache.get(f"kline:1min:sector:{primary_sector}") or []
     components: list[StockInfo] = cache.get(f"sector:components:{primary_sector}") or []
-    all_quotes = cache.get("quotes:batch") or []
-    quote_map = {q.code: q for q in all_quotes} if all_quotes else {}
+    quotes: list[Quote] = cache.get("quotes:batch") or []
+    qmap = {q.code: q for q in quotes}
 
-    if not stock_klines:
-        return ScoreResult(
-            dim="drive", score=30.0, weight=0.35,
-            details={"fallback": True, "reason": "无日K线数据"}
-        )
+    s_early, d_early = _early_seal(code, cache, primary_sector, components, qmap)
+    s_lead, d_lead = _lead_sector(minute, sector_min)
+    s_voice, d_voice = _voice(components, qmap)
 
-    # ─── Step 1: 找近 3 个涨停日 ───
-    limit_up_dates = _find_limit_up_dates(stock_klines, stock_1min)
-
-    if not limit_up_dates:
-        return ScoreResult(
-            dim="drive", score=30.0, weight=0.35,  # 兜底 30 分
-            details={"limit_up_count": 0, "reason": "近30日无涨停"}
-        )
-
-    # ─── 构建同板块候选股池 ───
-    peer_candidates = _build_peer_pool(code, primary_sector, candidate_pool, cache)
-
-    # ─── Step 2: 每个涨停日独立打分 ───
-    day_scores = []
-    day_details = []
-    for lu in limit_up_dates[:3]:  # 最多 3 天
-        ds, dd = _score_limit_up_day(
-            lu, primary_sector, components, quote_map, peer_candidates, cache
-        )
-        day_scores.append(ds)
-        day_details.append(dd)
-
-    # ─── Step 3: 取最佳天 ───
-    if len(day_scores) == 1:
-        best_idx = 0
-    else:
-        best_idx = max(range(len(day_scores)), key=lambda i: day_scores[i])
-
-    drive_score = day_scores[best_idx]
-    best_day_detail = day_details[best_idx]
-
-    # 连板加分
-    max_cons = max(lu["consecutive"] for lu in limit_up_dates)
-    if max_cons >= 2:
-        consecutive_bonus = min(max_cons * 5, 100 - drive_score)
-        drive_score += consecutive_bonus
-    else:
-        consecutive_bonus = 0
-
-    drive_score = min(drive_score, 100)
-
+    total = clip(s_early * R.EARLY_W + s_lead * R.LEAD_W + s_voice * R.VOICE_W)
     return ScoreResult(
-        dim="drive",
-        score=round(drive_score, 2),
-        weight=0.35,
-        details={
-            "limit_up_days": len(limit_up_dates[:3]),
-            "best_day": limit_up_dates[best_idx]["date"],
-            "best_day_detail": best_day_detail,
-            "consecutive_bonus": consecutive_bonus,
-            "max_consecutive": max_cons,
-        }
+        dim=DIM, score=round(total, 2), weight=WEIGHT,
+        details={"s_early": round(s_early, 2), "early": d_early,
+                 "s_lead": round(s_lead, 2), "lead": d_lead,
+                 "s_voice": round(s_voice, 2), "voice": d_voice},
     )
 
 
-# ─── 涨停日识别 ───
+# ─── 子因子①：封板最早（板块涨停封板池内相对排名）───
 
-def _find_limit_up_dates(day_klines: list[KBar], five_min_klines: list[KBar]) -> list[dict]:
-    """从日K线倒推近 3 个涨停日，附封板时间和换手率"""
-    limit_up_dates = []
-    consecutive = 0
+def _early_seal(code, cache, sector, components, qmap) -> tuple[float, dict]:
+    # 板块当日涨停股集合（含本股）
+    limit_codes = [s.code for s in components
+                   if s.code in qmap and qmap[s.code].pct >= R.LIMIT_UP_PCT]
+    if code not in limit_codes and qmap.get(code) and qmap[code].pct >= R.LIMIT_UP_PCT:
+        limit_codes.append(code)
+    pool = {}
+    for c in limit_codes:
+        mk: list[KBar] = cache.get(f"kline:1min:{c}") or []
+        q = qmap.get(c)
+        if mk and q and q.limit_up > 0:
+            t = _first_seal_minute(mk, q.limit_up)
+            if t is not None:
+                pool[c] = t
+    if code not in pool:
+        # 本股未封板（或无1分K）→ 0
+        q = qmap.get(code)
+        return 0.0, {"sealed": False, "pool_size": len(pool),
+                     "bid1_volume": q.bid1_volume if q else 0,
+                     "bid1_price": q.bid1_price if q else 0}
+    q = qmap.get(code)
+    seal_minute = pool[code]
+    detail = {"sealed": True, "seal_minute": seal_minute,
+              "seal_time": _fmt_minute_bucket(seal_minute),
+              "bid1_volume": q.bid1_volume if q else 0,
+              "bid1_price": q.bid1_price if q else 0,
+              "limit_up": q.limit_up if q else 0}
+    if len(pool) == 1:
+        return 100.0, {**detail, "rank": 1, "pool_size": 1}
+    # 封板时点升序排名（越早 rank 越小）
+    order = sorted(pool.items(), key=lambda kv: kv[1])
+    rank = [c for c, _ in order].index(code) + 1
+    n = len(pool)
+    s = (1.0 - rank / n) * 100.0
+    return clip(s), {**detail, "rank": rank, "pool_size": n}
 
-    for bar in reversed(day_klines):
-        if bar.pct >= 9.9:
-            consecutive += 1
-            # 从 5 分 K 找封板时间
-            board_time = _detect_board_time(five_min_klines, bar.timestamp)
-            date_str = datetime.fromtimestamp(bar.timestamp / 1000).strftime("%Y-%m-%d")
-            limit_up_dates.append({
-                "date": date_str,
-                "timestamp": bar.timestamp,
-                "board_time": board_time,       # "09:45" or None
-                "turnover": bar.turnover,
-                "consecutive": consecutive,
-                "pct": bar.pct,
-            })
-        else:
-            consecutive = 0
 
-    return limit_up_dates
-
-
-def _detect_board_time(five_min: list[KBar], day_ts: int) -> Optional[str]:
-    """从 5 分 K 线找封板时间（当天第一个接近涨停价的 bar）"""
-    if not five_min:
-        return None
-
-    day_date = datetime.fromtimestamp(day_ts / 1000).date()
-    day_bars = [b for b in five_min
-                if datetime.fromtimestamp(b.timestamp / 1000).date() == day_date]
-
-    if not day_bars:
-        return None
-
-    # 涨停价 ≈ 当天最高价（涨停日 high 就是涨停价）
-    limit_up_price = max(b.close for b in day_bars)
-    # 找到第一个 close 接近涨停价的 bar（0.1%误差容忍）
-    for bar in day_bars:
-        if limit_up_price > 0 and bar.close / limit_up_price >= 0.999:
-            dt = datetime.fromtimestamp(bar.timestamp / 1000)
-            return f"{dt.hour}:{dt.minute:02d}"
-
-    # fallback: close >= 当天第一根 bar open * 1.099
-    day_open = day_bars[0].open
-    for bar in day_bars:
-        if day_open > 0 and bar.close / day_open >= 1.099:
-            dt = datetime.fromtimestamp(bar.timestamp / 1000)
-            return f"{dt.hour}:{dt.minute:02d}"
-
+def _first_seal_minute(minute: list[KBar], limit_up: float) -> Optional[int]:
+    """首次封死涨停的分钟 bucket（close 到达涨停价）。"""
+    eps = limit_up * 0.001
+    for bar in minute:
+        if bar.close >= limit_up - eps:
+            return bar.timestamp // 60_000
     return None
 
 
-# ─── 同板块候选股池 ───
+def _fmt_minute_bucket(bucket: int) -> str:
+    return datetime.fromtimestamp(bucket * 60).strftime("%H:%M")
 
-def _build_peer_pool(code: str, primary_sector: str,
-                     candidate_pool: Optional[list[Candidate]],
-                     cache: DataCache) -> list[dict]:
-    """构建同板块候选股封板信息列表（供比对封板时间）"""
-    peers = []
-    if not candidate_pool:
-        # fallback: 从缓存推断
-        return peers
 
-    for cand in candidate_pool:
-        if cand.code == code:
+# ─── 子因子②：带动板块（脉冲-跟随因果检测）───
+
+def _lead_sector(stock: list[KBar], sector: list[KBar]) -> tuple[float, dict]:
+    if not stock or not sector:
+        return 40.0, {"degraded": True, "reason": "个股或板块1分K缺失"}
+    axis = common_minute_axis(stock, sector)
+    g_s = gain_curve(stock, axis)
+    g_b = gain_curve(sector, axis)
+    n = len(axis)
+    w = R.THRUST_WIN
+    L = R.LEAD_FOLLOW_BARS
+    thrust = R.THRUST_PCT / 100.0
+    follow_th = R.SECTOR_FOLLOW_PCT / 100.0
+
+    # ① 识别个股拉升脉冲起点 t0
+    thrusts: list[dict] = []
+    # 开盘跳空/一字高开
+    if g_s[0] is not None and g_s[0] >= thrust:
+        thrusts.append({"start": 0, "trigger": 0})
+    last_peak = -10
+    for t in range(w, n):
+        if g_s[t] is None or g_s[t - w] is None:
             continue
-        if primary_sector not in cand.concepts and not _is_same_sector(cand, primary_sector, cache):
+        dh = g_s[t] - g_s[t - w]
+        if dh >= thrust and (t - last_peak) > w:
+            thrusts.append({"start": t - w, "trigger": t})
+            last_peak = t
+
+    n_lead = 0
+    n_follow = 0
+    lead_events = []
+    for thrust_event in thrusts:
+        t0 = thrust_event["start"]
+        if g_b[t0] is None:
             continue
+        # ② 板块跟随振幅
+        seg = [g_b[t0 + k] for k in range(1, L + 1) if t0 + k < n and g_b[t0 + k] is not None]
+        if not seg:
+            continue
+        dh_b = max(seg) - g_b[t0]
+        # ③ 因果：跟随达标 + 板块未抢跑
+        pre = [g_b[t0 - k] for k in range(1, L + 1) if t0 - k >= 0 and g_b[t0 - k] is not None]
+        no_frontrun = (g_b[t0] - min(pre)) < follow_th if pre else True
+        if dh_b >= follow_th and no_frontrun:
+            n_lead += 1
+            lead_events.append(_build_lead_event(axis, g_s, g_b, t0,
+                                                 thrust_event.get("trigger", t0), L))
+        elif dh_b >= follow_th and not no_frontrun:
+            n_follow += 1
 
-        peer_1min = cache.get(f"kline:1min:{cand.code}") or []
-        peer_day = cache.get(f"kline:day:{cand.code}") or []
-        peer_lu = _find_limit_up_dates(peer_day, peer_1min)
+    # ④ 也检测「板块先达标、个股后拉」的被带动
+    follow_events = _detect_follow_events(axis, g_s, g_b, n, w, L, thrust, follow_th)
+    n_follow += len(follow_events)
 
-        # 取当天的封板时间
-        latest_date = datetime.now().strftime("%Y-%m-%d")
-        for lu in peer_lu:
-            if lu["date"] == latest_date and lu["board_time"]:
-                peers.append({
-                    "code": cand.code,
-                    "name": cand.name,
-                    "board_time": lu["board_time"],
-                    "board_timestamp": _time_to_minutes(lu["board_time"]),
-                })
-                break
-
-    return peers
-
-
-def _is_same_sector(cand: Candidate, sector_code: str, cache: DataCache) -> bool:
-    """检查候选股是否属于指定板块"""
-    components = cache.get(f"sector:components:{sector_code}") or []
-    for comp in components:
-        if comp.code == cand.code:
-            return True
-    return False
-
-
-def _time_to_minutes(t: str) -> int:
-    """"09:45" → 585 (分钟)"""
-    try:
-        h, m = t.split(":")
-        return int(h) * 60 + int(m)
-    except (ValueError, AttributeError):
-        return 9999  # 排序时放最后
-
-
-# ─── 单涨停日打分 ───
-
-def _score_limit_up_day(lu: dict, primary_sector: str,
-                         components: list[StockInfo],
-                         quote_map: dict,
-                         peer_candidates: list[dict],
-                         cache: DataCache) -> tuple[float, dict]:
-    """
-    Returns: (day_score, detail_dict)
-    """
-    # (a) 板块共鸣 Voice 30%
-    voice_score, voice_raw = _voice_score(primary_sector, components, quote_map)
-
-    # (b) 跟风力度 Follow 30%
-    follow_score, follow_raw = _follow_score(primary_sector, components, quote_map)
-
-    # (c) 封板决策力 Board Leadership 40%
-    board_score, board_detail = _board_leadership_score(
-        lu, peer_candidates, components, primary_sector
-    )
-    # 加入板块涨停总数
-    board_detail["sector_limit_up_total"] = voice_raw.get("limit_up", 0)
-
-    day_score = voice_score * 0.3 + follow_score * 0.3 + board_score * 0.4
-
-    return day_score, {
-        "voice": round(voice_score, 2),
-        "voice_raw": voice_raw,
-        "follow": round(follow_score, 2),
-        "follow_raw": follow_raw,
-        "board_leadership": round(board_score, 2),
-        "board_detail": board_detail,
-    }
-
-
-# ─── 子因子 A: 板块共鸣 ───
-
-
-def _resolve_component_pct(comp: StockInfo, quote_map: dict) -> float:
-    pct_val = comp.pct
-    if pct_val == 0.0 and comp.code in quote_map:
-        pct_val = quote_map[comp.code].pct
-    return pct_val
-
-
-def _active_components(components: list[StockInfo], quote_map: dict,
-                       limit: int = DRIVE_SAMPLE_LIMIT) -> list[StockInfo]:
-    ranked = sorted(
-        components,
-        key=lambda comp: (_resolve_component_pct(comp, quote_map), comp.code),
-        reverse=True,
-    )
-    return ranked[:limit]
-
-def _voice_score(sector_code: str, components: list[StockInfo],
-                 quote_map: dict) -> tuple[float, dict]:
-    """同行业涨停家数占比 → (score, raw_counts)"""
-    total = len(components)
-    scoring_components = _active_components(components, quote_map)
-    scoring_total = len(scoring_components)
-    limit_up_codes = []
-    for comp in scoring_components:
-        pct_val = _resolve_component_pct(comp, quote_map)
-        if pct_val >= 9.9:
-            limit_up_codes.append(comp.code)
-
-    limit_up_count = len(limit_up_codes)
-    if scoring_total == 0:
-        return 0.0, {"total": total, "scoring_total": 0, "sample_limit": DRIVE_SAMPLE_LIMIT, "limit_up": 0}
-
-    ratio = limit_up_count / scoring_total
-    score = min(ratio / 0.10, 1.0) * 100
-    return score, {
-        "total": total,
-        "scoring_total": scoring_total,
-        "sample_limit": DRIVE_SAMPLE_LIMIT,
-        "limit_up": limit_up_count,
-    }
-
-
-# ─── 子因子 B: 跟风力度 ───
-
-def _follow_score(sector_code: str, components: list[StockInfo],
-                  quote_map: dict) -> tuple[float, dict]:
-    """涨幅 >3% 但未涨停的占比 → (score, raw_counts)"""
-    total = len(components)
-    scoring_components = _active_components(components, quote_map)
-    scoring_total = len(scoring_components)
-    limit_up_codes = set()
-    strong_count = 0
-    down_count = 0
-
-    for comp in scoring_components:
-        pct_val = _resolve_component_pct(comp, quote_map)
-        if pct_val >= 9.9:
-            limit_up_codes.add(comp.code)
-        elif pct_val > 3.0:
-            strong_count += 1
-        if pct_val < 0:
-            down_count += 1
-
-    non_limit_total = scoring_total - len(limit_up_codes)
-    if non_limit_total == 0:
-        score = 100.0 if scoring_total > 0 else 0.0
-        return score, {
-            "total": total,
-            "scoring_total": scoring_total,
-            "sample_limit": DRIVE_SAMPLE_LIMIT,
-            "strong": strong_count,
-            "down": down_count,
-            "limit_up": len(limit_up_codes),
-        }
-
-    ratio = strong_count / non_limit_total
-    score = min(ratio / 0.15, 1.0) * 100
-    return score, {
-        "total": total,
-        "scoring_total": scoring_total,
-        "sample_limit": DRIVE_SAMPLE_LIMIT,
-        "strong": strong_count,
-        "down": down_count,
-        "limit_up": len(limit_up_codes),
-    }
-
-
-# ─── 子因子 C: 封板决策力 ───
-
-def _board_leadership_score(lu: dict, peers: list[dict],
-                            components: list[StockInfo],
-                            sector_code: str) -> tuple[float, dict]:
-    """
-    四子因子: C1(排名25%) + C2(绝对时间25%) + C3(紧密度50%)
-    × 一字板惩罚
-    """
-    # C1: 封板排名
-    rank_score, rank = _seal_rank_score(lu, peers)
-
-    # C2: 绝对时间
-    early_score = _early_time_score(lu.get("board_time"))
-
-    # C3: 小弟紧密度
-    gap_score, gap_detail = _gap_score(lu, peers)
-
-    board_score = rank_score * 0.25 + early_score * 0.25 + gap_score * 0.50
-
-    # 一字板惩罚
-    is_yizi = _is_yiziban(lu)
-    penalty = 0.85 if is_yizi else 1.0
-    board_score *= penalty
-
-    return board_score, {
-        "rank_score": round(rank_score, 2),
-        "seal_rank": rank,
-        "early_score": round(early_score, 2),
-        "board_time": lu.get("board_time"),
-        "gap_score": round(gap_score, 2),
-        "gap_detail": gap_detail,
-        "is_yiziban": is_yizi,
-        "penalty": penalty,
-    }
-
-
-def _seal_rank_score(lu: dict, peers: list[dict]) -> tuple[float, int]:
-    """C1: 在同板块涨停候选股中按封板时间排名"""
-    if not lu.get("board_time"):
-        return 50.0, len(peers) + 1  # 一字板或无封板信号，排最后
-
-    my_minutes = _time_to_minutes(lu["board_time"])
-    all_times = [(p["code"], p["board_timestamp"]) for p in peers]
-
-    # 只有封板时间的参与排名
-    ranked = sorted([t for _, t in all_times if t < 9999] + [my_minutes])
-
-    total = len(ranked)
-    if total == 0:
-        rank = 1
+    # ⑤ 打分
+    if n_lead == 0:
+        s_lead = 0.0  # 纯跟风票，follow 不再扣分
     else:
-        rank = ranked.index(my_minutes) + 1
+        base = R.LEAD_BASE + (n_lead - 1) * R.LEAD_STEP
+        penalty = min(n_follow * R.FOLLOW_PENALTY, R.FOLLOW_PENALTY_CAP)
+        s_lead = clip(base - penalty)
 
-    return (1 - (rank - 1) / total) * 100, rank
+    # ⑥ 整体方向 bonus（时滞互相关）
+    bonus = _corr_bonus(g_s, g_b, L)
+    if n_lead >= 1 and bonus > 0:
+        s_lead = clip(s_lead + bonus)
 
-
-def _early_time_score(board_time: Optional[str]) -> float:
-    """C2: 离散阶梯封板时间分"""
-    if not board_time:
-        return 1.0  # 一字板或无时间
-
-    h = int(board_time.split(":")[0])
-    m = int(board_time.split(":")[1])
-    minutes = h * 60 + m
-
-    if minutes <= 570:   # ≤9:30
-        return 100.0
-    elif minutes <= 630:  # ≤10:30
-        return 70.0
-    elif minutes <= 690:  # ≤11:30
-        return 40.0
-    else:
-        return 10.0         # 午后
+    return s_lead, {"n_lead": n_lead, "n_follow": n_follow,
+                    "n_thrust": len(thrusts), "bonus": bonus,
+                    "suspect_follower": n_lead == 0 and n_follow > 0,
+                    "lead_events": lead_events[:3],
+                    "follow_events": follow_events[:3]}
 
 
-def _gap_score(lu: dict, peers: list[dict]) -> tuple[float, dict]:
-    """C3: 小弟紧密度 — 其他涨停股与候选股封板时间差"""
-    if not lu.get("board_time"):
-        return 50.0, {"avg_gap_min": None, "within_5min_pct": 0, "peer_count": 0}
-
-    my_minutes = _time_to_minutes(lu["board_time"])
-    gaps = []
-
-    for p in peers:
-        if p["board_timestamp"] < 9999:  # 有封板时间的
-            gaps.append(abs(p["board_timestamp"] - my_minutes))
-
-    if not gaps:
-        # 独苗，无小弟
-        return 50.0, {"avg_gap_min": None, "within_5min_pct": 0, "peer_count": 0, "solo": True}
-
-    avg_gap = sum(gaps) / len(gaps)
-    within_5 = sum(1 for g in gaps if g <= 5) / len(gaps)
-
-    if avg_gap <= 5 and within_5 > 0.5:
-        score = 100.0
-    else:
-        score = max(0.0, 100 - avg_gap / 30 * 100)
-
-    return score, {
-        "avg_gap_min": round(avg_gap, 1),
-        "within_5min_pct": round(within_5, 2),
-        "peer_count": len(gaps),
+def _build_lead_event(axis, g_s, g_b, t0: int, trigger: int, L: int) -> dict:
+    """构造“个股先拉、板块跟随”的报告事件。"""
+    end = min(len(axis), t0 + L + 1)
+    sector_points = [(i, g_b[i]) for i in range(t0 + 1, end) if g_b[i] is not None]
+    stock_points = [(i, g_s[i]) for i in range(t0, end) if g_s[i] is not None]
+    sector_peak_idx, sector_peak = max(sector_points, key=lambda kv: kv[1]) if sector_points else (t0, g_b[t0])
+    stock_peak_idx, stock_peak = max(stock_points, key=lambda kv: kv[1]) if stock_points else (trigger, g_s[trigger])
+    stock_base = g_s[t0] or 0.0
+    sector_base = g_b[t0] or 0.0
+    return {
+        "event_time": _fmt_minute_bucket(axis[t0]),
+        "trigger_time": _fmt_minute_bucket(axis[trigger]),
+        "stock_peak_time": _fmt_minute_bucket(axis[stock_peak_idx]),
+        "sector_peak_time": _fmt_minute_bucket(axis[sector_peak_idx]),
+        "stock_gain_pct": round(((stock_peak or stock_base) - stock_base) * 100, 2),
+        "sector_gain_pct": round(((sector_peak or sector_base) - sector_base) * 100, 2),
     }
 
 
-def _is_yiziban(lu: dict) -> bool:
-    """一字板判定：无封板时间(board_time=None) 且 换手率 < 1%"""
-    if lu.get("board_time") is not None:
-        return False
-    return lu.get("turnover", 0) < 1.0
+def _detect_follow_events(axis, g_s, g_b, n, w, L, thrust, follow_th) -> list[dict]:
+    """板块先拉、个股随后跟（个股被带动）。"""
+    events = []
+    last = -10
+    for t in range(w, n):
+        if g_b[t] is None or g_b[t - w] is None:
+            continue
+        if (g_b[t] - g_b[t - w]) < follow_th or (t - last) <= w:
+            continue
+        t0 = t - w
+        seg = [g_s[t0 + k] for k in range(1, L + 1)
+               if t0 + k < n and g_s[t0 + k] is not None]
+        if seg and g_s[t0] is not None and (max(seg) - g_s[t0]) >= thrust:
+            stock_peak_offset, stock_peak = max(enumerate(seg, start=1), key=lambda kv: kv[1])
+            stock_follow_idx = min(t0 + stock_peak_offset, n - 1)
+            events.append({
+                "sector_event_time": _fmt_minute_bucket(axis[t0]),
+                "sector_trigger_time": _fmt_minute_bucket(axis[t]),
+                "stock_follow_time": _fmt_minute_bucket(axis[stock_follow_idx]),
+                "sector_gain_pct": round((g_b[t] - g_b[t0]) * 100, 2),
+                "stock_gain_pct": round((stock_peak - g_s[t0]) * 100, 2),
+            })
+            last = t
+    return events
+
+
+def _corr_bonus(g_s, g_b, L) -> float:
+    """时滞互相关：个股领先板块 τ*>0 且 ρ≥CORR_TH → +CORR_BONUS。"""
+    rs = _returns(g_s)
+    rb = _returns(g_b)
+    if len(rs) < 5:
+        return 0.0
+    best_tau, best_rho = 0, -2.0
+    for tau in range(0, L + 1):
+        rho = _corr(rs, rb, tau)
+        if rho is not None and rho > best_rho:
+            best_rho, best_tau = rho, tau
+    if best_tau > 0 and best_rho >= R.CORR_TH:
+        return R.CORR_BONUS
+    return 0.0
+
+
+def _returns(g) -> list[Optional[float]]:
+    out = [None]
+    for t in range(1, len(g)):
+        if g[t] is None or g[t - 1] is None:
+            out.append(None)
+        else:
+            out.append(g[t] - g[t - 1])
+    return out
+
+
+def _corr(rs, rb, tau) -> Optional[float]:
+    xs, ys = [], []
+    for t in range(len(rs)):
+        if t + tau >= len(rb):
+            break
+        a, b = rs[t], rb[t + tau]
+        if a is None or b is None:
+            continue
+        xs.append(a); ys.append(b)
+    if len(xs) < 5:
+        return None
+    mx = sum(xs) / len(xs); my = sum(ys) / len(ys)
+    cov = sum((x - mx) * (y - my) for x, y in zip(xs, ys))
+    vx = sum((x - mx) ** 2 for x in xs) ** 0.5
+    vy = sum((y - my) ** 2 for y in ys) ** 0.5
+    if vx <= EPS or vy <= EPS:
+        return None
+    return cov / (vx * vy)
+
+
+# ─── 子因子③：板块共鸣 ───
+
+def _voice(components: list[StockInfo], qmap) -> tuple[float, dict]:
+    actives = [s for s in components if s.code in qmap]
+    n = len(actives)
+    if n == 0:
+        return 0.0, {"degraded": True, "n": 0}
+    n_limit = sum(1 for s in actives if qmap[s.code].pct >= R.LIMIT_UP_PCT)
+    n_strong = sum(1 for s in actives if qmap[s.code].pct >= R.FOLLOW_PCT)
+    limit_ratio = n_limit / n
+    strong_ratio = n_strong / n
+    s_limit = clip(limit_ratio / R.VOICE_FULL, 0, 1) * 100.0
+    s_strong = clip(strong_ratio / R.FOLLOW_FULL, 0, 1) * 100.0
+    s_voice = s_limit * R.VOICE_LIMIT_W + s_strong * R.VOICE_STRONG_W
+    return clip(s_voice), {"n": n, "n_limit": n_limit, "n_strong": n_strong,
+                           "limit_ratio": round(limit_ratio, 3),
+                           "strong_ratio": round(strong_ratio, 3)}
