+++ dragon_quant/scorers/registry.py
@@ -1,4 +1,4 @@
-"""scorers_v2.registry —「识别真龙」五维评分体系的权重/门槛/阈值常量。
+"""scorers.registry —「识别真龙」五维评分体系的权重/门槛/阈值常量。
 
 全部集中于此，便于回测调参；算法文件只引用，不写死。
 依据《评分器Refactor.md》§9。
