+++ review/技术方案.md
@@ -1,19 +1,19 @@
 # Review 模块技术方案
 
-> 龙头回测验证系统 — 验证四维评分的实际表现
+> 龙头回测验证系统 — 验证五维评分的实际表现
 
 ---
 
 ## 一、背景
 
-`dragon-quant` 每天 scan 产出 topN 龙头的四维评分，写入 `dragons` 表。但这些龙头在后续交易日实际表现如何？评分高的票是否真的有更大机会？需要一套回测机制来量化验证。
+`dragon-quant` 每天 scan 产出 topN 龙头的五维评分，写入 `dragons_v2` 表。但这些龙头在后续交易日实际表现如何？评分高的票是否真的有更大机会？需要一套回测机制来量化验证。旧 `dragons_v1` 表仅作为历史记录显式查询。
 
 ## 二、核心逻辑
 
 ```
 scan 执行时                       review 执行时（用户手动触发）
 ───────                          ──────────
-写入 dragons                     读取 pending 记录
+写入 dragons_v2                  读取 pending 记录
 (trade_date + 基本面快照)          ↓
       ↓                          拉取历史日K（雪球）
 5 日窗口去重                      找入选日后第一个非一字板日 → 买入价
@@ -21,7 +21,7 @@ scan 执行时                       review 执行时（用户手动触发）
 buy_date = NULL                   ↓
 review_status = 'pending'         再算买入日至峰值日窗口内最大回撤
                                    ↓
-                                  UPDATE dragons
+                                  UPDATE dragons_v2
 ```
 
 ### 2.1 入选去重
